requirements.txt
openai>=0.44.1
numpy>=1.25.2
scikit-learn==1.3.1
umap-learn==0.5.4
pandas==2.1.1
path==16.7.1
sentence-transformers==2.2.2
spacy==3.7.0
hdbscan==0.8.33
bertopic>=0.15.0
tqdm==4.66.1
langchain-community>=0.1.0
langchain-openai>=0.1.0
tiktoken>=0.5.1
mecab-python3>=1.0.5
nltk>=3.8.1
huggingface_hub>=0.16.4
